# Conflicts:
#	js/react_native/android/gradle.properties
#	js/react_native/e2e/android/build.gradle
#	tools/ci_build/github/android/default_qnn_aar_build_settings.json
#	tools/ci_build/github/android/training_full_aar_build_settings.json
#	tools/ci_build/github/js/react_native_e2e_full_aar_build_settings.json

+++ onnxruntime/core/optimizer/attention_fusion_helper.h
@@ -281,8 +281,8 @@ bool ValidateUnidirMask(std::vector<T> mask_data, int64_t w, bool& is_undirectio
     is_undirectional = true;
 
     const T* p = mask_data.data();
-    for (int i = 0; i < w; i++) {
-      for (int j = 0; j < w; j++) {
+    for (int64_t i = 0; i < w; i++) {
+      for (int64_t j = 0; j < w; j++) {
         if (*p != static_cast<T>(1)) {
           is_one = false;
         }
